ref: final inner eval loop hooks (#3154)
* final inner eval loop hooks

* final inner eval loop hooks
